==> Ev Fiyat Tahmini/1-VeriCekme.ipynb <==
# kutuphaneler
from glob import glob
from os import link
import requests
from parsel import Selector
import pandas as pd
from fake_useragent import UserAgent
from sympy import re

ua = UserAgent()
header = {"User-Agent": str(ua.chrome)}
# ilan ozelliklerinin kaydedildigi degisken
ilanlar = []

# linklerin kaydedildigi degisken
linkler = []

# 1 sayfadan baslayarak x sayfa kadar tarar, sitede en son 1400 sayfa vardır
# yeni sonuc gelmeyince dongu durdurulur
for i in range(1, 1400):
    resp = requests.get(f"https://www.hepsiemlak.com/istanbul-satilik?page={i}", headers=header)
    selectorLink = Selector(resp.text)
    # x sayfadaki linkler alinir
    tempLink = selectorLink.xpath('//div[@class="links"]/a/@href').getall()
    # yeni sonuc gelmediyse dongu kirilir
    if len(tempLink) < 1:
        pass
    # sonuc varsa linklere eklenir
    linkler += tempLink
    print(f"{i}.sayfa bitti")

print(len(linkler))
# ilanlar birer birer dongu ile gezilir
def tara(linkler):
    
    hataliLinkler = []
    global ilanlar
    if input("Devam et? -1") is "1":
        print("Devam ediliyor...")
    else:
        return
    counter = 1
    for linkIndex in range(len(linkler)):
        ua = UserAgent()
        header = {"User-Agent": str(ua.chrome)}
        print(f"{counter}/{len(linkler)}---{len(ilanlar)}")
        counter+=1
        try:
            resp = requests.get(f"https://www.hepsiemlak.com{linkler[linkIndex]}",headers=header)
            selector = Selector(resp.text)
            # fiyat bilgileri alinir ve gereksiz bosluklar strip ile temizlenir
            fiyat = selector.xpath('//p[@class="fontRB fz24 price"]/text()').get().strip()
            #  il ilce mahalle bilgileri alinir ve gereksiz bosluklar temizlenir
            kisaDetay = [i.replace(" ","").replace("\n","").replace("m","") for i in selector.xpath('//ul[@class="short-info-list"]/li/text()').getall()]
            il = kisaDetay[0]
            ilce = kisaDetay[1]
            mahalle = kisaDetay[2]
            # detayli ilan bilgileri alinir
            detaylar = [i.strip().replace("\n","") for i in selector.xpath('//ul[@class="adv-info-list"]/li/span/text()').getall()]
            # veriler alinirken net ve brut m2 ayri ayri gorunuyor bu veriler bulunur ve listede birlestirilir (cekilen veri ile ilgili)
            for i in detaylar:
                if i[0] == "/":
                    birlestir = detaylar.index(i)
                    break
            detaylar[birlestir-1] += detaylar[birlestir]
            detaylar.pop(birlestir)
            # detaylar sutun ve icerik bilgisi olarak ayrilir sirayla tek cift durumuna gore
            sutunlar = [detaylar[i] for i in range(len(detaylar)) if i%2==0]
            icerik = [detaylar[i] for i in range(len(detaylar)) if i%2==1]

            # ilan bilgileri tempilan degiskenine atilir ve eksik veri olmadigi durumda genel ilanlar listesine atilir
            tempilan = [fiyat, il, ilce, mahalle]
            # sutun yani ozelliklerde dongu ile gezinilir ve istenen bilgiler ozelliklere eklenir
            for i in range(len(sutunlar)):
                if sutunlar[i] == "İlan Durumu":
                    tempilan.append(icerik[i])

                if sutunlar[i] == "Konut Şekli":
                    tempilan.append(icerik[i])
                    
                if sutunlar[i] == "Oda + Salon Sayısı":
                    tempilan.append(icerik[i])
                
                if sutunlar[i] == "Brüt / Net M2":
                    a,b = icerik[i].split("/ ")
                    tempilan.append(a)
                    tempilan.append(b)
                
                if sutunlar[i] == "Bulunduğu Kat":
                    tempilan.append(icerik[i])
                
                if sutunlar[i] == "Bina Yaşı":
                    tempilan.append(icerik[i])
                
                if sutunlar[i] == "Banyo Sayısı":
                    tempilan.append(icerik[i])
                
                if sutunlar[i] == "Yapı Tipi":
                    tempilan.append(icerik[i])
                
                if sutunlar[i] == "Eşya Durumu":
                    tempilan.append(icerik[i])
                
                if sutunlar[i] == "Site İçerisinde":
                    tempilan.append(icerik[i])
                
                if sutunlar[i] == "Yakıt Tipi":
                    tempilan.append(icerik[i])
            # ozelliklerin butunlugu kontrol edilir 16 ozelligin hepsi varsa genel listeye eklenir
            if len(tempilan) == 16:
                ilanlar.append(tempilan)
                sonuc = pd.DataFrame(data= ilanlar)
                #print(sonuc)
                # ilgili dosyalara donusturulur cikti alinir
                sonuc.to_csv("outputCSV.csv")
                sonuc.to_excel("outputExcel.xlsx")
        except Exception as e:
            print("HATA: ", e)
            hataliLinkler.append(linkler[linkIndex])
    return tara(linkler=hataliLinkler)

tara(linkler=linkler)
# dataframe olusturulur
sonuc = pd.DataFrame(data= ilanlar)
print(sonuc)
# ilgili dosyalara donusturulur cikti alinir
sonuc.to_csv("outputCSV.csv")
sonuc.to_excel("outputExcel.xlsx")


import pandas as pd

df = pd.read_csv("outputCSV.csv")
df.drop("Unnamed: 0", axis = 1, inplace = True)

# Sütun Adlarını Düzenleme
df_2 = df.rename(columns = {"0":"fiyat",
                     "1":"il",
                     "2":"ilce",
                     "3":"mahalle",
                     "4":"ilan_durumu",
                     "5":"konut",
                     "6":"oda_sayisi",
                     "7":"net_mt",
                     "8":"brut_mt",
                     "9":"dairenin_katı",
                     "10":"yas",
                     "11":"esya_durumu",
                     "12":"banyo_sayisi",
                     "13":"yapi_tipi",
                     "14":"site_icerisinde",
                     "15":"yakit_tipi"})

df_2.head()

df_2.info()

df_2.describe()
