==> Ev Fiyat Tahmini/3-Veri İşleme.ipynb <==
import pandas as pd

# veri seti okunur
df = pd.read_csv("outputCSV.csv")
df.drop("Unnamed: 0", axis = 1, inplace = True)

df.head()

# kolon isimleri adlandirilir
df_2 = df.rename(columns = {"0":"fiyat",
                     "1":"il",
                     "2":"ilce",
                     "3":"mahalle",
                     "4":"ilan_durumu",
                     "5":"konut",
                     "6":"oda_sayisi",
                     "7":"net_mt",
                     "8":"brut_mt",
                     "9":"dairenin_katı",
                     "10":"yas",
                     "11":"esya_durumu",
                     "12":"banyo_sayisi",
                     "13":"yapi_tipi",
                     "14":"site_icerisinde",
                     "15":"yakit_tipi"})

df_2.head()

df_2.columns

# veriler duzenlenir
df_2.loc[df_2.dairenin_katı == "Kot 3", "dairenin_katı"] = "-3"

df_2.site_icerisinde.value_counts()

df_2.head()

df_2.drop("site_icerisinde", axis = 1, inplace = True)
df_2.to_excel("yeni.xlsx")

df_2.head()

#Df setimizde boş değer var mı kontrolü
df.columns[df.isnull().any()]


from sklearn import preprocessing 
le = preprocessing.LabelEncoder()

df_2.columns

df = pd.read_excel("yeni.xlsx")
df.drop("Unnamed: 0", axis = 1, inplace = True)
df_2 = df.copy()
df_3 = df.copy()

df_3.head()

df_3["ilce"] = le.fit_transform(df_2.ilce)

le.classes_

df_3.ilce.unique()

le.inverse_transform([1])

Veri işlemede string değerler sayısal değerlere dönüştürülüyor 


df_3["konut"] = le.fit_transform(df_2.konut)

df_3["oda_sayisi"] = le.fit_transform(df_2.oda_sayisi)

df_3["dairenin_katı"] = le.fit_transform(df_2.dairenin_katı)

df_3["yas"] = le.fit_transform(df_2.yas)

df_3["esya_durumu"] = le.fit_transform(df_2.esya_durumu)

df_3["yapi_tipi"] = le.fit_transform(df_2.yapi_tipi)

df_3["yakit_tipi"] = le.fit_transform(df_2.yakit_tipi)

df_3.to_excel("yeni3.xlsx")

df_3.head()

df_3["banyo_sayisi"] = le.fit_transform(df_2.banyo_sayisi)

# hatali kisim duzelttik
df_3["fiyat"] = df_3["fiyat"].astype("int64")




df_3.head()

df_3.to_excel("yeni3.xlsx")
