==> Ev Fiyat Tahmini/4-Makine Öğrenmesinin Uygulanması (1).ipynb <==
## Gradient Boosting Regressor Algoritmasının Uygulanması

from sklearn import preprocessing 
le = preprocessing.LabelEncoder()
import numpy as np
from sklearn.model_selection import train_test_split, GridSearchCV,cross_val_score
import pandas as pd

df = pd.read_excel("yeni3.xlsx")
df.drop("Unnamed: 0", axis = 1, inplace = True)

df

Bazı değişkenlerin - değer çıkmasının sebebi fiyatlara negatif etki etmesindendir diye yorumladık.

print('Fiyat ile en iyi ilişkisi olan değişken')
corr = df.corr()
corr.sort_values(['fiyat'], ascending = False, inplace = True)
corr.fiyat

y = df['fiyat']

del df['fiyat']

X = df.values
y = y.values

#Verileri eğitim ve test verileri olmak üzere ikiye ayıralım (%60 eğitim, % 40 test)
from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.40, random_state = 42)

#modelimizi eğitelim
from sklearn.ensemble import GradientBoostingRegressor
GBR = GradientBoostingRegressor(n_estimators = 100, max_depth = 4)

#Veri setimize model uygulayalım
GBR.fit(X_train, y_train)

#Doğruluk oranımızı da kontrol edelim
print("Doğruluk Oranı :",(f"{ GBR.score(X_test, y_test)*100:.4f}"))

print(" Gerçek değerler:","\n",y_test[10:15])

print(" Tahmin edilen değerler:","\n", GBR.predict(X_test)[10:15])


print("Tahmin edilen değer : " + str(GBR.predict([X_test[9]])))
print("Gerçek değer : " + str(y_test[9]))

np.sqrt(-1*(cross_val_score(GBR, X_test, y_test, cv=10, scoring='neg_mean_squared_error'))).mean()
